+++ packages/plugin-hardhat/src/scripts/migrate-oz-cli-project.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { migrateLegacyProject } from '@openzeppelin/upgrades-core';
+import { migrateLegacyProject } from '@quai/quais-upgrades-core';
 
 migrateLegacyProject().catch(e => {
   console.error(e);