+++ packages/plugin-hardhat/src/utils/attach-abi.ts
@@ -1,10 +1,10 @@
 import { Contract, Signer } from 'ethers';
 
 import ITransparentUpgradeableProxyV5 from '../artifacts/ITransparentUpgradeableProxy.json';
-import ITransparentUpgradeableProxyV4 from '@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol/ITransparentUpgradeableProxy.json';
+import ITransparentUpgradeableProxyV4 from '@quai/quais-upgrades-core/artifacts/@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol/ITransparentUpgradeableProxy.json';
 
 import ProxyAdminV5 from '../artifacts/ProxyAdmin.json';
-import ProxyAdminV4 from '@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol/ProxyAdmin.json';
+import ProxyAdminV4 from '@quai/quais-upgrades-core/artifacts/@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol/ProxyAdmin.json';
 
 import { HardhatRuntimeEnvironment } from 'hardhat/types';