+++ packages/plugin-hardhat/src/utils/initializer-data.ts
@@ -1,4 +1,4 @@
-import { UpgradesError } from '@openzeppelin/upgrades-core';
+import { UpgradesError } from '@quai/quais-upgrades-core';
 import { Interface } from 'ethers';
 
 export function getInitializerData(