+++ packages/plugin-hardhat/src/utils/validate-impl.ts
@@ -8,7 +8,7 @@ import {
   Manifest,
   processProxyKind,
   ValidationOptions,
-} from '@openzeppelin/upgrades-core';
+} from '@quai/quais-upgrades-core';
 import { DeployData } from './deploy-impl';
 
 /**